+++ src-tauri/src/command/stockatking.rs
@@ -21,7 +21,7 @@ pub struct StocktakingRecordData {
     supply_name: String,
     unit_name: String,
     unit_price: u32,
-    quantity: u32,
+    quantity: f64,
 }
 
 #[derive(Debug, Clone, Deserialize)]
