+++ src-tauri/src/core/provided_ports/dto.rs
@@ -12,30 +12,30 @@ pub struct SupplierDTO {
     pub name: String,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct StocktakingDTO {
     pub id: String,
     pub stocktaken_date: i64,
     pub records: Vec<StocktakingRecordDTO>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct StocktakingRecordDTO {
     pub supply_id: String,
     pub supply_name: String,
     pub unit_name: String,
     pub unit_price: u32,
-    pub quantity: u32,
+    pub quantity: f64,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct JournalDTO {
     pub id: String,
     pub entry_date: i64,
     pub records: Vec<JournalRecordDTO>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct JournalRecordDTO {
     pub supply_id: String,
     pub supply_name: String,
@@ -43,7 +43,7 @@ pub struct JournalRecordDTO {
     pub supplier_name: String,
     pub unit_name: String,
     pub unit_price: u32,
-    pub quantity: u32,
+    pub quantity: f64,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -84,13 +84,13 @@ pub struct SearchStocktakingQuery {
     pub period_end: Option<i64>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct RecordStocktakingCommand {
     pub stocktaken_date: i64,
     pub records: Vec<StocktakingRecordDTO>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct EditStocktakingCommand {
     pub stocktaking_id: String,
     pub records: Vec<StocktakingRecordDTO>,
@@ -104,13 +104,13 @@ pub struct SearchJournalsQuery {
     pub supply_name: Option<String>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct RecordJournalCommand {
     pub entry_date: i64,
     pub records: Vec<JournalRecordDTO>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct EditJournalCommand {
     pub journal_id: String,
     pub records: Vec<JournalRecordDTO>,
